+++ airline-gateway/src/main/java/app/airlinegateway/AirlineGatewayApplication.java
@@ -1,7 +1,10 @@
 package app.airlinegateway;
 
+import app.security.JwtProviderLite;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication
 public class AirlineGatewayApplication {
@@ -10,4 +13,8 @@ public class AirlineGatewayApplication {
         SpringApplication.run(AirlineGatewayApplication.class, args);
     }
 
+    @Bean
+    JwtProviderLite getJwtProvider(@Value("${jwt.secret.access}") String token) {
+        return new JwtProviderLite(token);
+    }
 }