+++ airline-project/src/main/java/app/controllers/api/rest/AccountRestApi.java
@@ -1,27 +1,16 @@
 package app.controllers.api.rest;
 
 import app.dto.AccountDTO;
-import app.entities.account.Role;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
+import app.entities.Role;
+import io.swagger.annotations.*;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.data.domain.Page;
-
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 import javax.mail.MethodNotSupportedException;
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
@@ -56,7 +45,7 @@ public interface AccountRestApi {
     @GetMapping("/auth")
     @ApiOperation(value = "Get authenticated Account")
     @ApiResponses(value = {@ApiResponse(code = 200, message = "Account found")})
-    ResponseEntity<AccountDTO> getAuthenticatedAccount();
+    ResponseEntity<AccountDTO> getAuthenticatedAccount(HttpServletRequest request);
 
     @PostMapping
     @ApiOperation(value = "Create Account")