+++ airline-project/src/main/java/app/controllers/rest/AccountRestController.java
@@ -1,19 +1,23 @@
 package app.controllers.rest;
 
 import app.controllers.api.rest.AccountRestApi;
-import app.dto.AccountDTO;
-import app.entities.account.Role;
+import app.mappers.AccountMapper;
+import app.mappers.AircraftMapper;
+import app.security.JwtProviderLite;
 import app.services.interfaces.AccountService;
 import app.services.interfaces.RoleService;
+import app.dto.AccountDTO;
+import app.entities.Role;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.mapstruct.factory.Mappers;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.mail.MethodNotSupportedException;
+import javax.servlet.http.HttpServletRequest;
 import java.util.Collections;
 import java.util.List;
 
@@ -24,6 +28,9 @@ public class AccountRestController implements AccountRestApi {
 
     private final AccountService accountService;
     private final RoleService roleService;
+    private final JwtProviderLite jwtProvider;
+
+    private final AccountMapper accountMapper = Mappers.getMapper(AccountMapper.class);
 
     @Override
     public ResponseEntity<Page> getAllAccountsPages(Integer page, Integer size) {
@@ -40,29 +47,34 @@ public class AccountRestController implements AccountRestApi {
         var account = accountService.getAccountById(id);
         return account.isEmpty()
                 ? new ResponseEntity<>(HttpStatus.NOT_FOUND)
-                : new ResponseEntity<>(new AccountDTO(account.get()), HttpStatus.OK);
+                : new ResponseEntity<>(accountMapper.convertToAccountDTO(account.get()), HttpStatus.OK);
     }
 
     @Override
-    public ResponseEntity<AccountDTO> getAuthenticatedAccount() {
+    public ResponseEntity<AccountDTO> getAuthenticatedAccount(HttpServletRequest request) {
         log.info("getAuthenticatedAccount: get currently authenticated Account");
-        var authAccount = accountService.getAccountByEmail(
-                SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString());
-        return ResponseEntity.ok(new AccountDTO(authAccount));
+        String token = request.getHeader("Authorization");
+        if (token != null && token.startsWith("Bearer ")) {
+            token = token.substring(7);
+            String username = jwtProvider.extractClaims(token).get().getSubject();
+            var authAccount = accountService.getAccountByEmail(username);
+            return ResponseEntity.ok(accountMapper.convertToAccountDTO(authAccount));
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @Override
     public ResponseEntity<AccountDTO> createAccountDTO(AccountDTO accountDTO)
             throws MethodNotSupportedException {
         log.info("create: create new Account with email={}", accountDTO.getEmail());
-        return ResponseEntity.ok(new AccountDTO(accountService.saveAccount(accountDTO)));
+        return ResponseEntity.ok(accountMapper.convertToAccountDTO((accountService.saveAccount(accountDTO))));
     }
 
     @Override
     public ResponseEntity<AccountDTO> updateAccountDTOById(Long id, AccountDTO accountDTO)
             throws MethodNotSupportedException {
         log.info("update: update Account with id = {}", id);
-        return new ResponseEntity<>(new AccountDTO( accountService.updateAccount(id,accountDTO)), HttpStatus.OK);
+        return new ResponseEntity<>(accountMapper.convertToAccountDTO( accountService.updateAccount(id,accountDTO)), HttpStatus.OK);
     }
 
     @Override