+++ airline-project/src/main/java/app/repositories/RoleRepository.java
@@ -1,6 +1,6 @@
 package app.repositories;
 
-import app.entities.account.Role;
+import app.entities.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;