+++ airline-project/src/main/java/app/services/interfaces/AccountService.java
@@ -1,7 +1,7 @@
 package app.services.interfaces;
 
 import app.dto.AccountDTO;
-import app.entities.account.Account;
+import app.entities.Account;
 import org.springframework.data.domain.Page;
 
 import java.util.Optional;
@@ -18,5 +18,4 @@ public interface AccountService {
     Account getAccountByEmail(String email);
 
     void deleteAccountById(Long id);
-
 }