+++ airline-security/src/main/java/app/controllers/domain/JwtRequest.java
@@ -1,0 +1,12 @@
+package app.controllers.domain;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+public class JwtRequest {
+
+    private String username;
+    private String password;
+}