+++ airline-security/src/main/java/app/controllers/domain/RefreshJwtRequest.java
@@ -1,0 +1,11 @@
+package app.controllers.domain;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class RefreshJwtRequest {
+
+    public String refreshToken;
+}