+++ airline-security/src/main/java/app/repositories/AccountRepository.java
@@ -1,0 +1,17 @@
+package app.repositories;
+
+import app.account.Account;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AccountRepository extends JpaRepository<Account, Long> {
+
+    Page<Account> findAll(Pageable pageable);
+
+    Account getAccountByEmail(String email);
+
+    Account getAccountById(Long id);
+}