+++ airline-security/src/main/java/app/services/UserDetailsServiceImpl.java
@@ -1,0 +1,21 @@
+package app.services;
+
+import app.config.AccountDetailsImpl;
+import app.repositories.AccountRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserDetailsServiceImpl implements UserDetailsService {
+
+    private final AccountRepository repository;
+
+    @Override
+    public UserDetails loadUserByUsername(String s) throws UsernameNotFoundException {
+        return new AccountDetailsImpl(repository.getAccountByEmail(s));
+    }
+}