+++ airline-security/src/main/java/app/services/interfaces/AccountService.java
@@ -1,0 +1,7 @@
+package app.services.interfaces;
+
+import app.account.Account;
+
+public interface AccountService {
+    Account getAccountByEmail(String email);
+}