+++ common/src/main/java/app/security/JwtProvider.java
@@ -1,0 +1,13 @@
+package app.security;
+
+import io.jsonwebtoken.Claims;
+import org.springframework.lang.NonNull;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface JwtProvider {
+    Optional<Claims> extractClaims(String token);
+
+    Optional<List<String>> extractRoles(@NonNull String token);
+}